+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # master
 
+- Automatic conversion of custom scalars! Heavily inspired by the upcoming `graphql_ppx` release which also has this, it's now possible to have ReasonRelay automatically convert your custom scalars at runtime. More info in the docs for custom scalars.
+
 # 0.9.1
 
 - Fix bug with `bs-platform >7.3.0`.